Level editor save changes directry in levels, refactor of levels
