+++ js/levels/index.js
@@ -1,0 +1,6 @@
+import {Level1} from './logic/level01'
+import {Level2} from './logic/level02'
+import {Level3} from './logic/level03'
+import {Level4} from './logic/level04'
+
+export{ Level1, Level2, Level3, Level4 }
