+++ js/levels/logic/level01.js
@@ -1,0 +1,13 @@
+import {GameState} from '../../gameState'
+import {LevelPrototype} from './levelPrototype'
+let gState = new GameState().state;
+let _ = require( "lodash" );
+
+let lvl = require( "./../structures/level01.js" );
+export class Level1 extends LevelPrototype {
+	constructor() {
+		super();
+		let protoLevel = _.cloneDeep( this.prototypeLevel );
+		this.levelObj = _.merge( protoLevel, lvl );
+	}
+}
