+++ js/levels/logic/level04.js
@@ -1,0 +1,13 @@
+import {GameState} from '../../gameState'
+import {LevelPrototype} from './levelPrototype'
+let gState = new GameState().state;
+let _ = require( "lodash" );
+
+let lvl = require( "./../structures/level04.js" );
+export class Level4 extends LevelPrototype {
+	constructor() {
+		super();
+		let protoLevel = _.cloneDeep( this.prototypeLevel );
+		this.levelObj = _.merge( protoLevel, lvl );
+	}
+}
