+++ js/levels/structures/level01.js
@@ -1,0 +1,119 @@
+module.exports = {
+    "world": {
+        "bounds": {
+            "x1": 0,
+            "x2": 885,
+            "y1": 0,
+            "y2": 376
+        },
+        "sprites": [
+            {
+                "x": 0,
+                "y": 0,
+                "type": "game-background"
+            },
+            {
+                "x": 640,
+                "y": 0,
+                "type": "game-background"
+            }
+        ]
+    },
+    "platforms": [
+        {
+            "x": 0,
+            "y": 300,
+            "type": "platform",
+            "id": 1420213760305
+        },
+        {
+            "x": 197,
+            "y": 300,
+            "type": "platform2",
+            "id": 217827147784
+        },
+        {
+            "x": 506,
+            "y": 300,
+            "type": "platform",
+            "id": 7434560914
+        },
+        {
+            "x": 646,
+            "y": 300,
+            "type": "platform",
+            "id": 676154916637
+        },
+        {
+            "x": 646,
+            "y": 112,
+            "type": "tower1",
+            "id": 1359876869802
+        }
+    ],
+    "redSlimes": [
+        {
+            "x": 670,
+            "y": 10,
+            "type": "monster2",
+            "id": 1102978552889
+        },
+        {
+            "x": 216,
+            "y": 276,
+            "type": "monster2"
+        }
+    ],
+    "fallers": [
+        {
+            "x": 101,
+            "y": 190,
+            "type": "faller",
+            "id": 639706769795
+        }
+    ],
+    "slowFallers": [],
+    "trampolines": [
+        {
+            "x": 600,
+            "y": 270,
+            "type": "trampoline",
+            "id": 509219385629
+        },
+        {
+            "x": 65,
+            "y": 288,
+            "type": "trampoline"
+        }
+    ],
+    "lava": [
+        {
+            "x": 141,
+            "y": 332,
+            "type": "lava",
+            "id": 405688441655
+        },
+        {
+            "x": 254,
+            "y": 332,
+            "type": "lava2",
+            "id": 1044302674827
+        },
+        {
+            "x": 700,
+            "y": 332,
+            "type": "lava2",
+            "id": 16463500511
+        }
+    ],
+    "tnt": [],
+    "switchFallers": [],
+    "riders": [],
+    "arrows": [
+        {
+            "x": 91,
+            "y": 278,
+            "type": "arrow"
+        }
+    ]
+}
