+++ js/levels/structures/level04.js
@@ -1,0 +1,38 @@
+module.exports = {
+    "world": {
+        "bounds": {
+            "x1": 0,
+            "x2": 885,
+            "y1": 0,
+            "y2": 376
+        },
+        "sprites": [
+            {
+                "x": 0,
+                "y": 0,
+                "type": "game-background"
+            },
+            {
+                "x": 640,
+                "y": 0,
+                "type": "game-background"
+            }
+        ]
+    },
+    "platforms": [
+        {
+            "x": -9,
+            "y": 347,
+            "type": "platform"
+        }
+    ],
+    "redSlimes": [],
+    "fallers": [],
+    "slowFallers": [],
+    "trampolines": [],
+    "lava": [],
+    "tnt": [],
+    "switchFallers": [],
+    "riders": [],
+    "arrows": []
+}
