+++ server.js
@@ -34,7 +34,7 @@ request.on('data', function(chunk) {
   return 0;
 }
 console.log( filePath );
-  fs.writeFile( filePath, JSON.stringify( content, null, 4 ), function(err) {
+  fs.writeFile( 'js/levels/structures/' + filePath, 'module.exports = ' + JSON.stringify( content, null, 4 ), function(err) {
     if(err) {
         response.writeHead(500, {"Content-Type": "application/json", "Access-Control-Allow-Origin" : "*"});
         response.write(  JSON.stringify({
